General Clean Up and Destroy Handling
This is just some more general clean up of the names used.
Also fixed some minor potential cases where destruction of objects could lead to an exception when completely exiting a scene (if there were many generic objects spawned).
