--- testproject/Assets/ManualTests/Scripts/SoftSynchSwitchScene.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a52acfcd370f67d4688ffb4787bb8c5b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
