+++ testproject/Assets/Materials/BlueMaterial.mat.meta
