+++ testproject/Assets/Prefabs/SpawnObjectBlueVariant.prefab
