+++ testproject/Assets/Prefabs/SpawnObjectBlueVariant.prefab.meta
