+++ testproject/Assets/Prefabs/SpawnObjectOrangeVariant.prefab
