+++ testproject/Assets/Prefabs/SpawnObjectOrangeVariant.prefab.meta
